+++ pkg/staticserver/helpers.go
@@ -1,18 +1,14 @@
-package cdn
+package staticserver
 
-import (
-	"net"
-)
+import "net"
 
 func GetFreePort() (int, error) {
 	addr, err := net.ResolveTCPAddr("tcp", "127.0.0.1:0")
-
 	if err != nil {
 		return 0, err
 	}
 
 	listener, err := net.ListenTCP("tcp", addr)
-
 	if err != nil {
 		return 0, err
 	}
@@ -20,18 +16,4 @@ func GetFreePort() (int, error) {
 	defer listener.Close()
 
 	return listener.Addr().(*net.TCPAddr).Port, nil
-}
-
-func GetOutboundIP() (string, error) {
-	conn, err := net.Dial("udp", "8.8.8.8:80")
-
-	if err != nil {
-		return "", err
-	}
-
-	defer conn.Close()
-
-	localAddr := conn.LocalAddr().(*net.UDPAddr)
-
-	return localAddr.IP.String(), nil
 }